i64 support
Added `format` parameter after the following lines in swagger.yaml:

- contract/ldeps line 162
- commandreport/ldeps line 1347
- executionreport/ldeps line 1489
- GetProductFeeParams line 9198
- SyncRequest line 11137

In SessionCache and CachedEntitySet

- change all refs from `int` to `long`

+++ Tradovate.Samples/Accounting.cs
@@ -13,7 +13,7 @@ namespace Tradovate
 {
     class Accounting
     {
-        public static Account GetAccount(int myUserId)
+        public static Account GetAccount(long myUserId)
         {
             var apiInstance = new AccountingApi();
             var result = apiInstance.AccountList();
